+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2023-06-15
+### New
+- added `Position.toJson` and `Position.fromJson` (#82)
+- added `Position.speedFrom` to be able to calculate speed between two coordinates using timestamps (#82)
+
 ## [1.3.3] - 2023-06-08
 ### Fixed
 - added Access Offer Type filtering to the call event channel. (#81)